refactor: domain-jpa 모듈 단일화
