--- modules/infrastructure/common-domain-jpa/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'com.whoz_in'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.3.4'
-
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-test {
-    useJUnitPlatform()
-}
