+++ modules/infrastructure/domain-jpa/build.gradle
@@ -12,9 +12,6 @@ repositories {
 dependencies {
     compileOnly project(':modules:domain')
 
-    // 강한 의존성 결합
-    compileOnly project(':modules:infrastructure:common-domain-jpa')
-
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.3.4'
     runtimeOnly 'com.mysql:mysql-connector-j:8.3.0'
