+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/device/DeviceEntity.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import com.whoz_in.common_domain_jpa.BaseEntity;
+import com.whoz_in.domain_jpa.shared.BaseEntity;
 
 @Entity
 public class DeviceEntity extends BaseEntity {
