+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/device/DeviceJpaAdapter.java
@@ -11,6 +11,6 @@ public class DeviceJpaAdapter implements DeviceRepository {
     private final DeviceJpaRepository repository;
     @Override
     public void save(Device device) {
-
+        repository.save(new DeviceEntity());
     }
 }
