+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/member/AuthEntity.java
@@ -1,6 +1,6 @@
 package com.whoz_in.domain_jpa.member;
 
-import com.whoz_in.common_domain_jpa.BaseEntity;
+import com.whoz_in.domain_jpa.shared.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
