+++ modules/infrastructure/domain-jpa/src/main/java/com/whoz_in/domain_jpa/member/MemberEntity.java
@@ -8,7 +8,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import com.whoz_in.common_domain_jpa.BaseEntity;
+import com.whoz_in.domain_jpa.shared.BaseEntity;
 
 
 @Entity
