--- modules/infrastructure/domain-log-jpa/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'com.whoz_in'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compileOnly project(':modules:infrastructure:domain-jpa')
-
-    // 강한 의존성 결합
-    compileOnly project(':modules:infrastructure:common-domain-jpa')
-
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.3.4'
-    runtimeOnly 'com.mysql:mysql-connector-j:8.3.0'
-}
-
-test {
-    useJUnitPlatform()
-}
