--- modules/infrastructure/domain-log-jpa/src/main/java/com/whoz_in/domain_log_jpa/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.whoz_in.domain_log_jpa.config;
-
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EntityScan(basePackages = "com.whoz_in.common_domain_jpa")
-public class JpaConfig {
-}
