+++ modules/infrastructure/log-writer/build.gradle
@@ -10,14 +10,10 @@ repositories {
 }
 
 dependencies {
-    compileOnly project(":modules:infrastructure:domain-log-jpa")
-    compileOnly project(":modules:infrastructure:domain-jpa")
-    compileOnly project(":modules:infrastructure:common-domain-jpa")
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.3.4'
+    runtimeOnly 'com.mysql:mysql-connector-j:8.3.0'
 
-
-    // 컴파일시에만 의존
     compileOnly 'jakarta.annotation:jakarta.annotation-api:2.1.1'
-    compileOnly 'org.springframework.boot:spring-boot-starter-data-jpa:3.3.4'
 
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
