+++ modules/infrastructure/log-writer/src/main/java/com/whoz_in/log_writer/infra/managed/ManagedLogConverter.java
@@ -2,8 +2,8 @@ package com.whoz_in.log_writer.infra.managed;
 
 import com.whoz_in.log_writer.infra.managed.arp.ArpLog;
 import com.whoz_in.log_writer.infra.managed.mdns.MdnsLog;
-import com.whoz_in.domain_log_jpa.managed.ManagedLog;
-import com.whoz_in.domain_log_jpa.managed.ManagedLog.LogId;
+import com.whoz_in.domain_jpa.managed.ManagedLog;
+import com.whoz_in.domain_jpa.managed.ManagedLog.LogId;
 import java.sql.Timestamp;
 
 public class ManagedLogConverter {
