+++ modules/infrastructure/log-writer/src/main/java/com/whoz_in/log_writer/infra/managed/mdns/MdnsLogWriter.java
@@ -2,7 +2,7 @@ package com.whoz_in.log_writer.infra.managed.mdns;
 
 import com.whoz_in.log_writer.infra.managed.ManagedLogConverter;
 import com.whoz_in.log_writer.infra.managed.ManagedLogDAO;
-import com.whoz_in.domain_log_jpa.managed.ManagedLog;
+import com.whoz_in.domain_jpa.managed.ManagedLog;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
